IBM-Swift/Kitura#656 Work in progress. Moving code from class to class
